+++ src/components/Buzz/buzz.utils.ts
@@ -2,12 +2,16 @@ import { useRouter } from 'next/router';
 import { trpc } from '~/utils/trpc';
 import { CreateBuzzSessionInput } from '~/server/schema/stripe.schema';
 import { getClientStripe } from '~/utils/get-client-stripe';
-import { showErrorNotification } from '~/utils/notifications';
+import { showErrorNotification, showSuccessNotification } from '~/utils/notifications';
+import { useState } from 'react';
 
-export const useQueryBuzzPackages = () => {
+export const useQueryBuzzPackages = ({ onPurchaseSuccess }: { onPurchaseSuccess?: () => void }) => {
   const router = useRouter();
+  const [processing, setProcessing] = useState<boolean>(false);
+  const queryUtils = trpc.useContext();
 
   const { data: packages = [], isLoading } = trpc.stripe.getBuzzPackages.useQuery();
+
   const createBuzzSessionMutation = trpc.stripe.createBuzzSession.useMutation({
     onSuccess: async ({ url, sessionId }) => {
       if (url) await router.push(url);
@@ -24,6 +28,27 @@ export const useQueryBuzzPackages = () => {
     },
   });
 
+  const { mutateAsync: completeStripeBuzzPurchaseMutation } =
+    trpc.buzz.completeStripeBuzzPurchase.useMutation({
+      async onSuccess() {
+        await queryUtils.buzz.getUserAccount.invalidate();
+        setProcessing(false);
+        showSuccessNotification({
+          title: 'Transaction completed successfully!',
+          message: 'Your Buzz has been added to your account.',
+        });
+        onPurchaseSuccess?.();
+      },
+      onError(error) {
+        showErrorNotification({
+          title: 'There was an error while attempting to purchase buzz. Please contact support.',
+          error: new Error(error.message),
+        });
+
+        setProcessing(false);
+      },
+    });
+
   const createCheckoutSession = (data: CreateBuzzSessionInput) => {
     return createBuzzSessionMutation.mutateAsync(data);
   };
@@ -32,6 +57,8 @@ export const useQueryBuzzPackages = () => {
     packages,
     isLoading,
     createCheckoutSession,
-    creatingSession: createBuzzSessionMutation.isLoading,
+    completeStripeBuzzPurchaseMutation,
+    processing,
+    setProcessing,
   };
 };
